analyzer/requirements.txt
connexion
flask
pyyaml
jsonschema
httpx
requests
apscheduler
logging
pytz
pykafka
setuptools
connexion[swagger-ui]
connexion[flask]
connexion[uvicorn]

